kairos_integration/requirements.txt
# Kairos Integration Framework Requirements

# Core dependencies
pyyaml>=6.0.1
selenium>=4.15.0
beautifulsoup4>=4.12.0

# Google OAuth authentication
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Async support
aiohttp>=3.8.0
aiofiles>=23.0.0

# Image processing for screenshots
Pillow>=10.0.0

# System monitoring
psutil>=5.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For enhanced performance
lxml>=4.9.0
requests>=2.31.0